requirements.txt
absl-py==2.1.0
addict==2.4.0
anyio==4.6.2.post1
asttokens==3.0.0
asyncio==3.4.3
attrs==24.2.0
basicsr==1.4.2
beautifulsoup4==4.12.3
bettercam==1.0.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.8
config-parser==0.0.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
DateTime==5.5
decorator==5.1.1
discordwebhook==1.0.3
dxcam==0.0.5
exceptiongroup==1.2.2
executing==2.1.0
fake-headers==1.0.2
fake-useragent==1.5.1
filelock==3.16.1
flake8==7.2.0
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
future==1.0.0
grpcio==1.67.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.0
humanfriendly==10.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
iniconfig==2.1.0
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
keyboard==0.13.5
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
onnx==1.17.0
onnxruntime-directml==1.20.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.6.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyperclip==1.9.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.0
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tb-nightly==2.19.0a20241109
tensorboard-data-server==0.7.2
tesseract==0.1.3
tifffile==2024.9.20
tls-client==1.0.1
tomli==2.0.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.1.7
urllib3==1.26.20
urllib3-secure-extra==0.1.0
uuid==1.30
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
yapf==0.40.2
zipp==3.20.2
zope.interface==7.2
